Reformat packages to separate logic and allow for easier testing
